vcf/header/record/value/map/format/definition/v4_5: Update local genotype posterior probabilities type
This was changed from an integer to a float in samtools/hts-specs#851.

+++ noodles-vcf/src/header/record/value/map/format/definition/v4_5.rs
@@ -82,7 +82,7 @@ pub(super) fn definition(key: &str) -> Option<(Number, Type, &'static str)> {
         )),
         key::LOCAL_GENOTYPE_POSTERIOR_PROBABILITIES => Some((
             Number::LocalSamples,
-            Type::Integer,
+            Type::Float,
             "Local-allele representation of GP",
         )),
         key::LOCAL_ROUNDED_GENOTYPE_LIKELIHOODS => Some((
